Ajustanto CRUD das entidades e organizando visualizações e dando forma ao formato do site

+++ Agenda_AspNet/Controllers/ContatoController.cs
@@ -7,16 +7,20 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using Agenda_AspNet.Data;
 using Agenda_AspNet.Models;
+using Microsoft.Extensions.Logging;
+using System.Diagnostics;
 
 namespace Agenda_AspNet.Controllers
 {
     public class ContatoController : Controller
     {
         private readonly Context _context;
+        private readonly ILogger<ContatoController> _logger;
 
-        public ContatoController(Context context)
+        public ContatoController(Context context, ILogger<ContatoController> logger)
         {
             _context = context;
+            _logger = logger;
         }
 
         // GET: Contato
@@ -40,6 +44,14 @@ namespace Agenda_AspNet.Controllers
                 return NotFound();
             }
 
+            var categoria = await _context.Categorias.FindAsync(contato.categoria_id);
+            contato.categoria = categoria;
+
+            List<Endereco> enderecos = await _context.Enderecos
+                .Where(x => x.contato_id == contato.id)
+                .ToListAsync();
+            contato.enderecos = enderecos;
+
             return View(contato);
         }
 
@@ -57,12 +69,19 @@ namespace Agenda_AspNet.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("nome,sobrenome,telefone,email,descricao,categoria_id,foto")] Contato contato)
         {
-            Console.WriteLine(contato.categoria_id);
             if (ModelState.IsValid)
             {
-                _context.Add(contato);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                try
+                {
+                    _context.Add(contato);
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
+                }
+                catch
+                {
+                    return NotFound();
+                }
+                
             }
             ViewBag.categoria_id = new SelectList(_context.Categorias, "id", "descricao", contato.categoria_id);
             return View(contato);
@@ -81,6 +100,7 @@ namespace Agenda_AspNet.Controllers
             {
                 return NotFound();
             }
+            ViewBag.categoria_id = new SelectList(_context.Categorias, "id", "descricao", contato.categoria_id);
             return View(contato);
         }
 
@@ -89,7 +109,7 @@ namespace Agenda_AspNet.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("id,nome,sobrenome,telefone,email,data_criacao,descricao,foto,ativo")] Contato contato)
+        public async Task<IActionResult> Edit(int id, [Bind("id,nome,sobrenome,telefone,email,data_criacao,descricao,categoria_id,foto,ativo")] Contato contato)
         {
             if (id != contato.id)
             {
@@ -116,6 +136,7 @@ namespace Agenda_AspNet.Controllers
                 }
                 return RedirectToAction(nameof(Index));
             }
+            ViewBag.categoria_id = new SelectList(_context.Categorias, "id", "descricao", contato.categoria_id);
             return View(contato);
         }
 
@@ -142,15 +163,26 @@ namespace Agenda_AspNet.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var contato = await _context.Contatos.FindAsync(id);
-            _context.Contatos.Remove(contato);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            if (ContatoExists(id))
+            {
+                var contato = await _context.Contatos.FindAsync(id);
+                _context.Contatos.Remove(contato);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+
+            return NotFound();
         }
 
         private bool ContatoExists(int id)
         {
             return _context.Contatos.Any(e => e.id == id);
+        }
+
+        [ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
+        public IActionResult Error()
+        {
+            return View(new ErrorViewModel { RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier });
         }
     }
 }
